put everything into a single file
as needed for integration into core one day
