+++ package/PKG_ADD
@@ -1,0 +1,4 @@
+autoload("h5read","h5read.oct")
+autoload("h5readatt","h5read.oct")
+autoload("h5write","h5read.oct")
+autoload("h5writeatt","h5read.oct")
